Commit of polling-error-resubscribe
